Bugfix/chatContainer

+++ src/views/ChatContainer.vue
@@ -8,7 +8,7 @@
         :render-label="renderLabel"
         :render-tag="renderMultipleSelectTag"
         filterable
-        placeholder="Add username"
+        placeholder="Добавить пользователей"
         type="text"
       />
       <button type="submit" :disabled="disableForm || !options">Создать новый чат</button>
@@ -23,11 +23,11 @@
         :render-label="renderLabel"
         :render-tag="renderMultipleSelectTag"
         filterable
-        placeholder="Add username"
+        placeholder="Добавить пользователей"
         type="text"
       />
-      <button type="submit" :disabled="disableForm || !options">Add User</button>
-      <button class="button-cancel" @click="inviteRoomId = null">Cancel</button>
+      <button type="submit" :disabled="disableForm || !options">Добавить пользователя</button>
+      <button class="button-cancel" @click="inviteRoomId = null">Отменить</button>
     </form>
 
     <form v-if="removeRoomId" @submit.prevent="deleteRoomUser">
@@ -37,8 +37,8 @@
           {{ user.username }}
         </option>
       </select>
-      <button type="submit" :disabled="disableForm || !removeUserId">Remove User</button>
-      <button class="button-cancel" @click="removeRoomId = null">Cancel</button>
+      <button type="submit" :disabled="disableForm || !removeUserId">Удалить пользователя</button>
+      <button class="button-cancel" @click="removeRoomId = null">Отменить</button>
     </form>
 
     <chat-window
@@ -133,14 +133,14 @@ export default {
       removeUserId: "",
       removeUsers: [],
       roomActions: [
-        { name: "inviteUser", title: "Invite User" },
-        { name: "removeUser", title: "Remove User" },
-        { name: "deleteRoom", title: "Delete Room" },
+        { name: "inviteUser", title: "Пригласить пользователя" },
+        { name: "removeUser", title: "Удалить пользователя" },
+        { name: "deleteRoom", title: "Удалить чат" },
       ],
       menuActions: [
-        { name: "inviteUser", title: "Invite User" },
-        { name: "removeUser", title: "Remove User" },
-        { name: "deleteRoom", title: "Delete Room" },
+        { name: "inviteUser", title: "Пригласить пользователя" },
+        { name: "removeUser", title: "Удалить пользователя" },
+        { name: "deleteRoom", title: "Удалить чат" },
       ],
       messageSelectionActions: [{ name: "deleteMessages", title: "Delete" }],
       styles: { container: { borderRadius: "4px" } },
@@ -203,7 +203,7 @@ export default {
               },
               [
                 h(NAvatar, {
-                  src: "https://07akioni.oss-cn-beijing.aliyuncs.com/07akioni.jpeg",
+                  src: "https://storage.googleapis.com/bairro-alto.appspot.com/queima-das-fitas32x32-1647209754638.webp",
                   round: true,
                   size: 22,
                   style: {
@@ -226,7 +226,7 @@ export default {
         },
         [
           h(NAvatar, {
-            src: "https://07akioni.oss-cn-beijing.aliyuncs.com/07akioni.jpeg",
+            src: "https://storage.googleapis.com/bairro-alto.appspot.com/queima-das-fitas32x32-1647209754638.webp",
             round: true,
             size: 24,
             style: {
@@ -563,6 +563,8 @@ export default {
         sender_id: this.currentUserId,
         content,
         timestamp: new Date(),
+        files: null,
+        replyMessage: null,
       };
 
       if (files) {
@@ -584,18 +586,14 @@ export default {
       const { id } = await firestoreService.addMessage(roomId, message);
 
       if (files) {
-        for (let index = 0; index < files.length; index += 1) {
-          // eslint-disable-next-line no-await-in-loop
-          await this.uploadFile({ file: files[index], messageId: id, roomId });
-        }
+        files.forEach((file) => this.uploadFile({ file, messageId: id, roomId }));
       }
 
       firestoreService.updateRoom(roomId, { lastUpdated: new Date() });
     },
 
     async editMessage({ messageId, newContent, roomId, files }) {
-      const newMessage = { edited: new Date() };
-      newMessage.content = newContent;
+      const newMessage = { edited: new Date(), content: newContent, files: null };
 
       if (files) {
         newMessage.files = this.formattedFiles(files);
@@ -606,12 +604,11 @@ export default {
       await firestoreService.updateMessage(roomId, messageId, newMessage);
 
       if (files) {
-        for (let index = 0; index < files.length; index += 1) {
-          if (files[index]?.blob) {
-            // eslint-disable-next-line no-await-in-loop
-            await this.uploadFile({ file: files[index], messageId, roomId });
+        files.forEach((file) => {
+          if (file?.blob) {
+            this.uploadFile({ file, messageId, roomId });
           }
-        }
+        });
       }
     },
 
@@ -769,15 +766,12 @@ export default {
     },
 
     messageSelectionActionHandler({ action, messages, roomId }) {
-      switch (action.name) {
-        case "deleteMessages":
-          return messages.forEach((message) => {
-            this.deleteMessage({ message, roomId });
-          });
-
-        default:
-          return false;
+      if (action.name === "deleteMessages") {
+        return messages.forEach((message) => {
+          this.deleteMessage({ message, roomId });
+        });
       }
+      return false;
     },
 
     async sendMessageReaction({ reaction, remove, messageId, roomId }) {
